Implement new event requestedNewGlyphCreate
